Improve condition

+++ src/Http/Controllers/Controller.php
@@ -141,13 +141,10 @@ abstract class Controller extends BaseController
                 $data->{$row->field} = str_replace($uuid, $data->getKey(), $data->{$row->field});
             });
             $data->save();
-            if ($old_path != $new_path) {
-                try {
-                    $request->session()->forget([$slug.'_path', $slug.'_uuid']);
-
-                    Storage::disk(config('voyager.storage.disk'))->move($old_path, $new_path);
-                    Storage::disk(config('voyager.storage.disk'))->deleteDirectory($folder_path);
-                } catch (Exception $e) { }
+            if ($old_path != $new_path && !Storage::disk(config('voyager.storage.disk'))->exists($new_path)) {
+                $request->session()->forget([$slug.'_path', $slug.'_uuid']);
+                Storage::disk(config('voyager.storage.disk'))->move($old_path, $new_path);
+                Storage::disk(config('voyager.storage.disk'))->deleteDirectory($folder_path);
             }
         }
